+++ PirateCore/src/main/java/lucatruglia/piratecore/managers/LevelManager.java
@@ -4,7 +4,7 @@ public class LevelManager {
     private static LevelManager instance;
 
     private double multiplier;
-    private long firstLevelXP;
+    private double firstLevelXP;
 
     public static LevelManager getInstance() {
         if (instance == null) {
@@ -15,7 +15,7 @@ public class LevelManager {
 
     public void initialize() {
         this.multiplier = ConfigManager.getInstance().getDouble("settings/levels.yml", "multiplier");
-        this.firstLevelXP = ConfigManager.getInstance().getInt("settings/levels.yml", "firstLevelXP");
+        this.firstLevelXP = ConfigManager.getInstance().getDouble("settings/levels.yml", "firstLevelXP");
 
         instance = this;
     }
@@ -24,48 +24,48 @@ public class LevelManager {
         return this.multiplier;
     }
 
-    public long getfirstLevelXP(){
+    public double getfirstLevelXP(){
         return this.firstLevelXP;
     }
 
-    public long getTotalXpNeededForLevel(int level) {
+    public double getTotalXpNeededForLevel(int level) {
         if (level <= 1)
             return 0;
 
-        long totalXp = 0;
-        long xpNeeded = firstLevelXP;
+        double totalXp = 0;
+        double xpNeeded = firstLevelXP;
 
         for (int i = 1; i < level; i++) {
             totalXp += xpNeeded;
-            xpNeeded = (long) (xpNeeded * multiplier);
+            xpNeeded = xpNeeded * multiplier;
         }
 
         return totalXp;
     }
 
-    public long getXpNeededForLevel(int level) {
+    public double getXpNeededForLevel(int level) {
         if (level <= 1)
             return 0;
 
-        long xpNeeded = firstLevelXP;
+        double xpNeeded = firstLevelXP;
         for (int i = 2; i < level; i++) {
-            xpNeeded = (long) (xpNeeded * multiplier);
+            xpNeeded = xpNeeded * multiplier;
         }
         return xpNeeded;
     }
 
-    public int getLevelByXP(long xp) {
+    public int getLevelByXP(double xp) {
         if (xp < firstLevelXP)
             return 1;
 
         int level = 1;
-        long totalXpNeeded = 0;
-        long xpNeeded = firstLevelXP;
+        double totalXpNeeded = 0;
+        double xpNeeded = firstLevelXP;
 
         while (xp >= totalXpNeeded + xpNeeded) {
             totalXpNeeded += xpNeeded;
             level++;
-            xpNeeded = (long) (xpNeeded * multiplier);
+            xpNeeded = xpNeeded * multiplier;
         }
 
         return level;
